VUSB: warning fix (comparison between signed and unsigned integer expressions)
svn:sync-xref-src-repo-rev: r124650

+++ src/VBox/Devices/USB/VUSBUrbPool.cpp
@@ -1,4 +1,4 @@
-/* $Id: VUSBUrbPool.cpp 73097 2018-07-12 21:06:33Z knut.osmundsen@oracle.com $ */
+/* $Id: VUSBUrbPool.cpp 73925 2018-08-28 09:33:32Z klaus.espenlaub@oracle.com $ */
 /** @file
  * Virtual USB - URB pool.
  */
@@ -129,7 +129,7 @@ DECLHIDDEN(PVUSBURB) vusbUrbPoolAlloc(PVUSBURBPOOL pUrbPool, VUSBXFERTYPE enmTyp
     /* Get the required amount of additional memory to allocate the whole state. */
     size_t cbMem = cbData + sizeof(VUSBURBVUSBINT) + cbHci + cTds * cbHciTd;
 
-    AssertReturn(enmType < RT_ELEMENTS(pUrbPool->aLstFreeUrbs), NULL);
+    AssertReturn((size_t)enmType < RT_ELEMENTS(pUrbPool->aLstFreeUrbs), NULL);
 
     RTCritSectEnter(&pUrbPool->CritSectPool);
     PVUSBURBHDR pHdr = NULL;
@@ -233,7 +233,7 @@ DECLHIDDEN(void) vusbUrbPoolFree(PVUSBURBPOOL pUrbPool, PVUSBURB pUrb)
     {
         /* Put it into the list of free URBs. */
         VUSBXFERTYPE enmType = pUrb->enmType;
-        AssertReturnVoid(enmType < RT_ELEMENTS(pUrbPool->aLstFreeUrbs));
+        AssertReturnVoid((size_t)enmType < RT_ELEMENTS(pUrbPool->aLstFreeUrbs));
         RTCritSectEnter(&pUrbPool->CritSectPool);
         pUrb->enmState = VUSBURBSTATE_FREE;
         RTListAppend(&pUrbPool->aLstFreeUrbs[enmType], &pHdr->NdFree);
